+++ packages/expo-ui/android/src/main/java/expo/modules/ui/TextInputView.kt
@@ -32,7 +32,9 @@ data class TextInputProps(
   val numberOfLines: MutableState<Int?> = mutableStateOf(null),
   val keyboardType: MutableState<String> = mutableStateOf("default"),
   val autocorrection: MutableState<Boolean> = mutableStateOf(true),
-  val autoCapitalize: MutableState<String> = mutableStateOf("none")
+  val autoCapitalize: MutableState<String> = mutableStateOf("none"),
+  val leadingIcon: MutableState<String?> = mutableStateOf(null),
+  val trailingIcon: MutableState<String?> = mutableStateOf(null)
 ) : ComposeProps
 
 private fun String.keyboardType(): KeyboardType {
@@ -83,7 +85,9 @@ class TextInputView(context: Context, appContext: AppContext) :
           keyboardType = props.keyboardType.value.keyboardType(),
           autoCorrectEnabled = props.autocorrection.value,
           capitalization = props.autoCapitalize.value.autoCapitalize()
-        )
+        ),
+        leadingIcon = props.leadingIcon.value?.let { { MaterialIcon(it) } },
+        trailingIcon = props.trailingIcon.value?.let { { MaterialIcon(it) } }
       )
     }
   }