+++ packages/expo-ui/CHANGELOG.md
@@ -15,6 +15,7 @@
 - [Android] Add `testID` support to compose primitives. ([#38005](https://github.com/expo/expo/pull/38005) by [@mateoguzmana](https://github.com/mateoguzmana))
 - [iOS] Add `testID` property to views. ([#37919](https://github.com/expo/expo/pull/37919) by [@mateoguzmana](https://github.com/mateoguzmana))
 - Add `SecureField` component. ([#37642](https://github.com/expo/expo/pull/37642) by [@aleqsio](https://github.com/aleqsio))
+- [Android] Add `leadingIcon` and `trailingIcon` properties to `TextInput`. ([#38147](https://github.com/expo/expo/pull/38147) by [@mateoguzmana](https://github.com/mateoguzmana))
 
 ### 🐛 Bug fixes